Add new component HistoricalView
